+++ CHANGELOG.md
@@ -38,6 +38,8 @@ been moved an renamed.
 - A float environment base class.
 - An unfinished Quantity class that can be used in conjunction with the
     quantitities package. https://pythonhosted.org/quantities/
+- Allow supplying a mapper function to dumps\_list and the add\_row method for
+    tabular like objects.
 
 ## [0.8.0] - 23-05-2015
 ### Added